test-containers in S3StoreTest
